# pkg/app/api/docs/swagger.yaml
basePath: /api/v1
definitions:
  ErrorResponse:
    properties:
      code:
        example: Internal
        type: string
      error:
        example: original server error message
        type: string
      msg:
        example: internal server error
        type: string
      status:
        example: error
        type: string
      trace_id:
        example: 1234567890abcdef1234567890abcdef
        type: string
    type: object
  SuccessResponse:
    properties:
      data:
        type: object
      status:
        example: success
        type: string
    type: object
...
  models.Agent:
    properties:
      auth_status:
        enum:
        - authorized
        - unauthorized
        - blocked
        type: string
      connected_date:
        type: string
      created_date:
        type: string
      deleted_at:
        type: string
      description:
        maxLength: 255
        type: string
      group_id:
        minimum: 0
        type: integer
      hash:
        type: string
      id:
        minimum: 0
        type: integer
      info:
        $ref: '#/definitions/models.AgentInfo'
      ip:
        maxLength: 50
        type: string
      status:
        enum:
        - connected
        - disconnected
        type: string
      updated_at:
        type: string
      version:
        maxLength: 20
        type: string
    required:
    - auth_status
    - description
    - hash
    - info
    - ip
    - status
    - version
    type: object
...
  models.AgentInfo:
    properties:
      net:
        $ref: '#/definitions/models.AgentNet'
      os:
        $ref: '#/definitions/models.AgentOS'
      tags:
        items:
          type: string
        maxItems: 20
        type: array
        uniqueItems: true
      users:
        items:
          $ref: '#/definitions/models.AgentUser'
        type: array
    required:
    - net
    - os
    - tags
    - users
    type: object
  models.AgentNet:
    properties:
      hostname:
        type: string
      ips:
        items:
          type: string
        type: array
    required:
    - hostname
    - ips
    type: object
  models.AgentOS:
    properties:
      arch:
        enum:
        - "386"
        - amd64
        type: string
      name:
        maxLength: 255
        type: string
      type:
        enum:
        - windows
        - linux
        - darwin
        type: string
    required:
    - arch
    - name
    - type
    type: object
  models.AgentUpgradeTask:
    properties:
      agent_id:
        minimum: 1
        type: integer
      batch:
        type: string
      created:
        type: string
      id:
        minimum: 0
        type: integer
      last_update:
        type: string
      reason:
        type: string
      status:
        enum:
        - new
        - running
        - ready
        - failed
        type: string
      version:
        maxLength: 20
        type: string
    required:
    - batch
    - status
    - version
    type: object
  models.AgentUser:
    properties:
      groups:
        items:
          type: string
        type: array
      name:
        type: string
    required:
    - groups
    - name
    type: object
...
  models.Group:
    properties:
      created_date:
        type: string
      deleted_at:
        type: string
      hash:
        type: string
      id:
        minimum: 0
        type: integer
      info:
        $ref: '#/definitions/models.GroupInfo'
      updated_at:
        type: string
    required:
    - hash
    - info
    type: object
...
  models.GroupInfo:
    properties:
      name:
        $ref: '#/definitions/models.GroupItemLocale'
      system:
        type: boolean
      tags:
        items:
          type: string
        maxItems: 20
        type: array
        uniqueItems: true
    required:
    - name
    - tags
    type: object
  models.GroupItemLocale:
    properties:
      en:
        maxLength: 4096
        type: string
      ru:
        maxLength: 4096
        type: string
    required:
    - en
    - ru
    type: object
...
  protected.upgradeAgent:
    properties:
      details:
        $ref: '#/definitions/protected.upgradeAgentDetails'
      task:
        $ref: '#/definitions/models.AgentUpgradeTask'
    type: object
  protected.upgradeAgentDetails:
    properties:
      agent:
        $ref: '#/definitions/models.Agent'
      group:
        $ref: '#/definitions/models.Group'
    type: object
  protected.upgradesAgents:
    properties:
      tasks:
        items:
          $ref: '#/definitions/models.AgentUpgradeTask'
        type: array
      total:
        type: integer
    type: object
  protected.upgradesAgentsAction:
    properties:
      filters:
        items:
          $ref: '#/definitions/storage.TableFilter'
        type: array
      version:
        type: string
    required:
    - version
    type: object
  protected.upgradesAgentsActionResult:
    properties:
      batch:
        type: string
      total:
        type: integer
    type: object
...
  storage.TableFilter:
    properties:
      field:
        type: string
      value:
        type: object
    required:
    - field
    - value
    type: object
info:
  contact:
    email: admin@vxcontrol.com
    name: Dmitry Nagibin
    url: https://vxcontrol.com
  description: Swagger API for VXControl SOLDR backend product.
  license:
    name: Proprietary License
    url: https://raw.githubusercontent.com/vxcontrol/soldr/master/LICENSE
  termsOfService: http://swagger.io/terms/
  title: SOLDR Swagger API
  version: "1.0"
paths:
...
  /upgrades/agents:
    get:
      parameters:
      - collectionFormat: multi
        description: |-
          Filtering result on server e.g. {"value":[...],"field":"..."}
            field value should be integer or string or array type
        in: query
        items:
          type: string
        name: filters[]
        type: array
      - description: Field to group results by
        in: query
        name: group
        type: string
      - default: en
        description: Language of result data
        enum:
        - en
        - ru
        in: query
        name: lang
        required: true
        type: string
      - default: 1
        description: Number of page (since 1)
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - default: 5
        description: Amount items per page (min -1, max 100, -1 means unlimited)
        in: query
        maximum: 100
        minimum: -1
        name: pageSize
        required: true
        type: integer
      - default: '{}'
        description: |-
          Sorting result on server e.g. {"prop":"...","order":"..."}
            field order is "ascending" or "descending" value
        in: query
        name: sort
        required: true
        type: string
      - default: init
        description: Type of request
        enum:
        - sort
        - filter
        - init
        - page
        - size
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: agents upgrades list received successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/protected.upgradesAgents'
              type: object
        "400":
          description: invalid query request data
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: getting agents upgrades not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on getting agents upgrades
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Retrieve agents upgrades list
      tags:
      - Upgrades
      - Agents
    post:
      consumes:
      - application/json
      parameters:
      - description: action on agents as JSON data
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/protected.upgradesAgentsAction'
      produces:
      - application/json
      responses:
        "201":
          description: agents upgrade requested succesful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/protected.upgradesAgentsActionResult'
              type: object
        "400":
          description: invalid agents upgrade request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: upgrading agents not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: agent binary file not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on requesting agents to upgrade
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Upgrade agents to a specific version
      tags:
      - Upgrades
      - Agents
  /upgrades/agents/{hash}/last:
    get:
      parameters:
      - description: agent hash in hex format (md5)
        in: path
        maxLength: 32
        minLength: 32
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: last agent upgrade information received successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/protected.upgradeAgent'
              type: object
        "400":
          description: invalid query request data
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: getting last agent upgrade information not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: agent or group or task not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on getting last agent upgrade information
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Retrieve last agent upgrade information
      tags:
      - Upgrades
      - Agents
    put:
      consumes:
      - application/json
      parameters:
      - description: agent hash in hex format (md5)
        in: path
        maxLength: 32
        minLength: 32
        name: hash
        required: true
        type: string
      - description: agent info as JSON data
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/models.AgentUpgradeTask'
      produces:
      - application/json
      responses:
        "200":
          description: last agent upgrade information updated successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.AgentUpgradeTask'
              type: object
        "400":
          description: invalid last agent upgrade information
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: updating last agent upgrade information not permitted
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: agent or group or task not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: internal error on updating last agent upgrade information
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Update last agent upgrade information
      tags:
      - Upgrades
      - Agents
...
swagger: "2.0"